requirements.txt
Django==2.2
django-expiring-token==1.0.2
django-otp==1.0.2
djangorestframework==3.9.4
mysqlclient==2.0.2
pkg-resources==0.0.0
pyotp==2.4.1
pytz==2020.4
sqlparse==0.4.1
Pillow==8.0.1
